refactor: move QPC_MIN_SERVER_VERSION to release module where it makes more logical sense to live
Relates to JIRA: DISCOVERY-1430
