+++ qpc/tests/report/test_report_download.py
@@ -11,10 +11,11 @@ import requests_mock
 
 from qpc import messages
 from qpc.cli import CLI
+from qpc.release import QPC_MIN_SERVER_VERSION
 from qpc.report import REPORT_URI
 from qpc.report.download import ReportDownloadCommand
 from qpc.scan import SCAN_JOB_URI
-from qpc.utils import QPC_MIN_SERVER_VERSION, create_tar_buffer, get_server_location
+from qpc.utils import create_tar_buffer, get_server_location
 
 
 @pytest.fixture
