+++ qpc/tests/report/test_report_show.py
@@ -11,10 +11,11 @@ import requests_mock
 
 from qpc import messages
 from qpc.cli import CLI
+from qpc.release import QPC_MIN_SERVER_VERSION
 from qpc.report import REPORT_V2_URI
 from qpc.report.show import ReportShowCommand
 from qpc.tests.utilities import redirect_stdout
-from qpc.utils import QPC_MIN_SERVER_VERSION, get_server_location
+from qpc.utils import get_server_location
 
 
 @pytest.fixture
