+++ qpc/utils.py
@@ -54,7 +54,6 @@ DEFAULT_INSIGHTS_CONFIG = {
 
 LOG_LEVEL_INFO = 0
 
-QPC_MIN_SERVER_VERSION = "2.7.0"
 
 logging.captureWarnings(True)
 logger = logging.getLogger(__name__)
